+++ gems/aws-sdk-core/CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased Changes
 ------------------
 
+* Issue - Convert stubs at request time.
+
 3.220.0 (2025-03-04)
 ------------------